backend/requirements.txt
django>=5.0
djangorestframework
django-cors-headers
python-dotenv
pymupdf
openai
requests
python-docx
